Main container
Added a div and put all my entire html code in there as I added a container to the div and with that on CSS added a margin of auto for left and right to make the webpage responsive. I also did the same for our CTA container to make it center in margin.