Gap and nested flexbox
Gap helps with spacing on flex item as it more convenient than margin so we added some gap to the flex items. And we also added some nested flexbox to the nav and footer to make things in our flex container have another flex container in which we can do other flex properties to it like make the footer items instead of being in columns, into rows.