Grid item and alignment
We added grid-row and grid-column so at our feature section, it alternates from text, photo, photo, text, text, photo. It better to do this instead of manually doing it because it always code quality to do text first and then images on grids. We also changed some color of the footer, and links at the bottom. We also changed the heart color at the footer by creating a span and having a class to change the color of the heart.