Breakpoint
Learned about break point in css. Create a new css files and this is normal when we want to accommodate for different screen sizes. So we just worked on things for each screen size to make sure our all screen sizes can have a good looking website (if you forgot about how to do this, watch the chapters again).