Positioning
We learn more about positioning and learn that its kinda a last resort when you want to layout things because you can just use flexblox for that. However things like have the navbar stay fixed when you scroll or having a chatbot icon consistent when you scroll are ways position can be used. So we added a chatbot icon on the botton of our page and use position to move it from the bottom left to the bottom right.